+++ innd/cc.c
@@ -1032,9 +1032,9 @@ CCnewgroup(char *av[])
 	if (*who == '\0')
 	    who = NEWSMASTER;
 
-        length = snprintf(NULL, 0, "%s %ld %s\n", Name, Now.time, who) + 1;
+        length = snprintf(NULL, 0, "%s %ld %s\n", Name, (long) Now.time, who) + 1;
         buff = xmalloc(length);
-        snprintf(buff, length, "%s %ld %s\n", Name, Now.time, who);
+        snprintf(buff, length, "%s %ld %s\n", Name, (long) Now.time, who);
 	if (xwrite(fd, buff, strlen(buff)) < 0) {
 	    oerrno = errno;
 	    syslog(L_ERROR, "%s cant write %s %m", LogName, TIMES);
